Dear PopV team,
We are following up on this issue: https://github.com/YosefLab/popV/issues/97
We observed a similar error than the authors of the issue, which we observed when launching the code on both a Jupyter Lab notebook and a bash script. We managed to patch it by adding a transfer to numpy array and re-converting afterwards to CSR. While this is modestly costly, and certainly brute-forcing, this offers a patch to this problem that at least three users have been facing.

Warm regards,
Eva and Soufiane